test: split MenuBar ITs into separate files, disable animations
