+++ vaadin-menu-bar-flow-parent/vaadin-menu-bar-flow-integration-tests/src/main/frontend/not-animated-styles.css
@@ -1,0 +1,6 @@
+:host([opening]),
+:host([closing]),
+:host([opening]) [part='overlay'],
+:host([closing]) [part='overlay'] {
+  animation: none !important;
+}
