+++ vaadin-menu-bar-flow-parent/vaadin-menu-bar-flow-integration-tests/src/test/java/com/vaadin/flow/component/menubar/tests/MenuBarPageIT.java
@@ -16,7 +16,6 @@
 package com.vaadin.flow.component.menubar.tests;
 
 import java.util.List;
-import java.util.Set;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -422,237 +421,6 @@ public class MenuBarPageIT extends AbstractComponentIT {
         verifyClosed();
     }
 
-    public void toggleMenuBarTheme_themeIsToggled() {
-        Assert.assertFalse(menuBar.hasAttribute("theme"));
-        click("toggle-theme");
-        Assert.assertEquals(menuBar.getDomAttribute("theme"),
-                MenuBarTestPage.MENU_BAR_THEME);
-        click("toggle-theme");
-        Assert.assertFalse(menuBar.hasAttribute("theme"));
-    }
-
-    @Test
-    public void toggleMenuItemTheme_themeIsToggled() {
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("theme"));
-        click("toggle-item-1-theme");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(menuButton1.getDomAttribute("theme"),
-                MenuBarTestPage.MENU_ITEM_THEME);
-        click("toggle-item-1-theme");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("theme"));
-    }
-
-    @Test
-    public void toggleMenuItemClassName_classNameIsToggled() {
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("class"));
-        click("toggle-item1-class-name");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(MenuBarTestPage.MENU_ITEM_FIRST_CLASS_NAME),
-                menuButton1.getClassNames());
-        click("toggle-item1-class-name");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(), menuButton1.getClassNames());
-    }
-
-    @Test
-    public void setMenuItemClassName_classNameIsSet() {
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("class"));
-        click("toggle-item1-class-name");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(MenuBarTestPage.MENU_ITEM_FIRST_CLASS_NAME),
-                menuButton1.getClassNames());
-        click("set-item1-class-name");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(MenuBarTestPage.MENU_ITEM_SECOND_CLASS_NAME),
-                menuButton1.getClassNames());
-    }
-
-    @Test
-    public void toggleMenuItemClassNameWithSetClassName_classNameIsToggled() {
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("class"));
-        click("set-unset-item1-class-name");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(MenuBarTestPage.MENU_ITEM_FIRST_CLASS_NAME),
-                menuButton1.getClassNames());
-        click("set-unset-item1-class-name");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(), menuButton1.getClassNames());
-    }
-
-    @Test
-    public void toggleMultipleItemClassName_classNamesAreToggled() {
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("class"));
-        click("add-remove-multiple-classes");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(
-                Set.of(MenuBarTestPage.MENU_ITEM_FIRST_CLASS_NAME,
-                        MenuBarTestPage.MENU_ITEM_SECOND_CLASS_NAME),
-                menuButton1.getClassNames());
-        click("add-remove-multiple-classes");
-        menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(Set.of(), menuButton1.getClassNames());
-    }
-
-    @Test
-    public void subMenuHasClassName_callRemoveClassName_classNameIsRemoved() {
-        verifySubMenuItemClassNames(true,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME);
-
-        click("remove-sub-item-class-name");
-        verifyClosed();
-        verifySubMenuItemClassNames(false,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME);
-    }
-
-    @Test
-    public void subMenuItem_toggleMultipleClassNames_classNamesAreToggled() {
-        click("add-second-sub-item-class-name");
-        verifyClosed();
-        verifySubMenuItemClassNames(true,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME,
-                MenuBarTestPage.SUB_ITEM_SECOND_CLASS_NAME);
-
-        click("add-remove-multiple-sub-item-classes");
-        verifyClosed();
-        verifySubMenuItemClassNames(false,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME,
-                MenuBarTestPage.SUB_ITEM_SECOND_CLASS_NAME);
-
-        click("add-remove-multiple-sub-item-classes");
-        verifyClosed();
-        verifySubMenuItemClassNames(true,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME,
-                MenuBarTestPage.SUB_ITEM_SECOND_CLASS_NAME);
-    }
-
-    @Test
-    public void subMenuItem_toggleSingleClassName_classNameIsToggled() {
-        click("toggle-sub-item-class-name");
-        verifyClosed();
-        verifySubMenuItemClassNames(false,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME);
-
-        click("toggle-sub-item-class-name");
-        verifyClosed();
-        verifySubMenuItemClassNames(true,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME);
-    }
-
-    @Test
-    public void subMenuItem_classNamesAreToggleWithSet_classNamesAreToggled() {
-        click("set-unset-sub-item-class-name");
-        verifyClosed();
-        verifySubMenuItemClassNames(false,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME);
-
-        click("set-unset-sub-item-class-name");
-        verifyClosed();
-        verifySubMenuItemClassNames(true,
-                MenuBarTestPage.SUB_ITEM_FIRST_CLASS_NAME);
-    }
-
-    @Test
-    public void menuItemWithClassNameInOverflow_changeClassName_classNameIsChanged() {
-        click("set-width");
-        click("set-item2-class-name");
-        waitForResizeObserver();
-        menuBar.getOverflowButton().click();
-        click("change-item2-class-name");
-        menuBar.getOverflowButton().click();
-        TestBenchElement menuItem = menuBar.getSubMenuItems().get(0);
-        Assert.assertEquals(Set.of(MenuBarTestPage.MENU_ITEM_SECOND_CLASS_NAME),
-                menuItem.getClassNames());
-    }
-
-    @Test
-    public void menuItemWithClassNameInOverflow_removeClassName_classNameIsRemoved() {
-        click("set-width");
-        click("set-item2-class-name");
-        waitForResizeObserver();
-        menuBar.getOverflowButton().click();
-        click("remove-item2-class-name");
-        menuBar.getOverflowButton().click();
-        TestBenchElement menuItem = menuBar.getSubMenuItems().get(0);
-
-        Assert.assertEquals(Set.of(), menuItem.getClassNames());
-    }
-
-    @Test
-    public void menuItemWithClassNameInOverflow_menuItemLeavesOverflow_classNameCanBeChanged() {
-        click("set-width");
-        click("set-item2-class-name");
-        waitForResizeObserver();
-        menuBar.getOverflowButton().click();
-        click("reset-width");
-        waitForResizeObserver();
-        click("change-item2-class-name");
-        TestBenchElement menuItem = menuBar.getButtons().get(1);
-        Assert.assertEquals(Set.of(MenuBarTestPage.MENU_ITEM_SECOND_CLASS_NAME),
-                menuItem.getClassNames());
-
-        click("remove-item2-class-name");
-        menuItem = menuBar.getButtons().get(1);
-        Assert.assertEquals(Set.of(), menuItem.getClassNames());
-    }
-
-    @Test
-    public void setMenuItemTheme_toggleVisibility_themeIsPreserved() {
-        click("toggle-item-1-theme");
-        click("toggle-item-1-visibility");
-        click("toggle-item-1-visibility");
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(menuButton1.getDomAttribute("theme"),
-                MenuBarTestPage.MENU_ITEM_THEME);
-    }
-
-    @Test
-    public void setMenuItemTheme_hide_resetTheme_show_themeIsUnset() {
-        click("toggle-item-1-theme");
-        click("toggle-item-1-visibility");
-        click("toggle-item-1-theme");
-        click("toggle-item-1-visibility");
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertFalse(menuButton1.hasAttribute("theme"));
-    }
-
-    @Test
-    public void toggleSubMenuItemTheme_themeIsToggled() {
-        menuBar.getButtons().get(0).click();
-        Assert.assertFalse(
-                menuBar.getSubMenuItems().get(1).hasAttribute("theme"));
-
-        click("toggle-sub-theme");
-        verifyClosed();
-
-        menuBar.getButtons().get(0).click();
-        Assert.assertEquals(
-                menuBar.getSubMenuItems().get(1).getDomAttribute("theme"),
-                MenuBarTestPage.SUB_ITEM_THEME);
-
-        click("toggle-sub-theme");
-        verifyClosed();
-
-        menuBar.getButtons().get(0).click();
-        Assert.assertFalse(
-                menuBar.getSubMenuItems().get(1).hasAttribute("theme"));
-    }
-
-    @Test
-    public void toggleMenuBarTheme_toggleMenuItemTheme_themeIsOverridden() {
-        click("toggle-theme");
-        click("toggle-item-1-theme");
-
-        TestBenchElement menuButton1 = menuBar.getButtons().get(0);
-        Assert.assertEquals(MenuBarTestPage.MENU_ITEM_THEME,
-                menuButton1.getDomAttribute("theme"));
-    }
-
     @After
     public void afterTest() {
         checkLogsForErrors();
@@ -735,20 +503,5 @@ public class MenuBarPageIT extends AbstractComponentIT {
 
     public void verifyOpened() {
         waitForElementPresent(By.tagName(OVERLAY_TAG));
-    }
-
-    private void verifySubMenuItemClassNames(boolean containsClassNames,
-            String... classNames) {
-        openSubSubMenu();
-        verifyOpened();
-        TestBenchElement subMenuItem = menuBar.getSubMenuItems().get(2);
-        var subMenuItemClassNames = subMenuItem.getClassNames();
-        for (String className : classNames) {
-            if (containsClassNames) {
-                Assert.assertTrue(subMenuItemClassNames.contains(className));
-            } else {
-                Assert.assertFalse(subMenuItemClassNames.contains(className));
-            }
-        }
     }
 }
